backend/requirements.txt
langgraph
langchain
langchain-openai
langchain-community
langchain-core
langchain-google-genai
langfuse
fastapi
uvicorn
python-dotenv
psycopg==3.2.9
psycopg-binary
psycopg-pool==3.2.6
redis
pgvector
sse-starlette
pre-commit
pandas
loguru
requests-oauthlib